requirements.txt
asgiref==3.10.0
Django==5.2.7
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
django-filter==24.3
drf-spectacular==0.27.2
django-cors-headers==4.6.0

pillow==11.3.0
PyJWT==2.10.1
python-dotenv==1.1.1
sqlparse==0.5.3
tzdata==2025.2

PyMySQL>=1.1          

reportlab